[backend] API addPirDependency
